+++ src/lifecycle.ts
@@ -1,8 +1,8 @@
 import assert from 'node:assert';
 import { EventEmitter } from 'node:events';
 import { debuglog } from 'node:util';
-import is, { isClass } from 'is-type-of';
-import ReadyObject from 'get-ready';
+import { isClass } from 'is-type-of';
+import { Ready as ReadyObject } from 'get-ready';
 import type { ReadyFunctionArg } from 'get-ready';
 import { Ready } from 'ready-callback';
 import { EggConsoleLogger } from 'egg-logger';
@@ -109,8 +109,13 @@ export class Lifecycle extends EventEmitter {
     });
   }
 
-  ready(arg?: ReadyFunctionArg) {
-    return this.#readyObject.ready(arg);
+  ready(): Promise<void>;
+  ready(flagOrFunction: ReadyFunctionArg): void;
+  ready(flagOrFunction?: ReadyFunctionArg) {
+    if (flagOrFunction === undefined) {
+      return this.#readyObject.ready();
+    }
+    return this.#readyObject.ready(flagOrFunction);
   }
 
   get app() {
@@ -184,7 +189,7 @@ export class Lifecycle extends EventEmitter {
   }
 
   registerBeforeClose(fn: Fun) {
-    assert(is.function(fn), 'argument should be function');
+    assert(typeof fn === 'function', 'argument should be function');
     assert(this.#isClosed === false, 'app has been closed');
     this.#closeFunctionSet.add(fn);
   }