+++ test/fixtures/plugin-pkg-exports/config/plugin.js
@@ -2,7 +2,4 @@ module.exports = {
   a: {
     enable: true,
   },
-  b: {
-    enable: true,
-  },
 };