+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [6.1.0](https://github.com/eggjs/egg-core/compare/v6.0.3...v6.1.0) (2024-12-17)
+
+
+### Features
+
+* support `pkg.eggPlugin.exports` property ([#274](https://github.com/eggjs/egg-core/issues/274)) ([df9efed](https://github.com/eggjs/egg-core/commit/df9efedb22c630b13a4f1374f96450f6ba1107ab))
+
+## [6.0.3](https://github.com/eggjs/egg-core/compare/v6.0.2...v6.0.3) (2024-12-13)
+
+
+### Bug Fixes
+
+* dont import default value ([#273](https://github.com/eggjs/egg-core/issues/273)) ([16274e7](https://github.com/eggjs/egg-core/commit/16274e744007bcbeef9a99210fa00c8812d05718))
+
 ## [6.0.2](https://github.com/eggjs/egg-core/compare/v6.0.1...v6.0.2) (2024-06-30)